feat(preview-tearsheet): implement open in stackblitz in web component
