+++ packages/ibm-products-web-components/src/components/tearsheet-preview/tearsheet.stories.ts
@@ -9,14 +9,9 @@
 
 import { html } from 'lit';
 import './index';
-import styles from './_story-assets/_storybook-styles.scss?lit';
-import { prefix } from '../../globals/settings';
+import styles from './story-styles.scss?lit';
 import './_story-assets/step-tearsheet-preview';
 import './_story-assets/stacking-tearsheet';
-import { iconLoader } from '@carbon/web-components/es/globals/internal/icon-loader.js';
-import Bee from '@carbon/icons/es/bee/32';
-import Add16 from '@carbon/icons/es/add/32';
-import RightPanelClose32 from '@carbon/icons/es/right-panel--close/32';
 import '@carbon/web-components/es/components/progress-indicator/index.js';
 import '@carbon/web-components/es/components/tabs/index.js';
 import '@carbon/web-components/es/components/ai-label/index.js';
@@ -27,270 +22,26 @@ import '@carbon/web-components/es/components/toggle-tip/index.js';
 import '@carbon/web-components/es/components/dropdown/index.js';
 import '@carbon/web-components/es/components/progress-bar/index.js';
 import '@carbon/web-components/es/components/notification/index.js';
-import { breakpoints } from '@carbon/layout';
 import mdx from './tearsheet.mdx';
-
-const storyPrefix = 'tearsheet-preview-stories';
-
-// Helper to get button size based on screen size and variant
-const getButtonSize = (variant: string = 'wide') => {
-  const smMediaQuery = `(max-width: ${breakpoints.md.width})`;
-  const isSmallScreen = window.matchMedia(smMediaQuery).matches;
-  return isSmallScreen || variant === 'narrow' ? 'xl' : '2xl';
-};
-
-// When decorator=true, render an AI label; false/undefined renders nothing.
-const getDecorator = (decorator) => {
-  if (!decorator) {
-    return '';
-  }
-  return html`
-    <cds-ai-label alignment="bottom-right" slot="decorator">
-      <div slot="body-text">
-        <p class="secondary">AI Explained</p>
-        <h2 class="ai-label-heading">84%</h2>
-        <p class="secondary bold">Confidence score</p>
-        <p class="secondary">Any description goes here</p>
-        <hr />
-        <p class="secondary">Model type</p>
-        <p class="bold">Foundation model</p>
-      </div>
-    </cds-ai-label>
-  `;
-};
-const toggleButton = () => {
-  document
-    .querySelector(`${prefix}-preview-tearsheet`)
-    ?.toggleAttribute('open');
-};
-
-/**
- * Handles `c4p-preview-tearsheet-collapse-change` events in stories.
- * Resizes header-action buttons to `xs` when collapsed, `sm` when expanded,
- * matching the React TearsheetHeaderActions behavior.
- */
-const handleCollapseChange = (e: Event) => {
-  const { collapsed } = (e as CustomEvent).detail as { collapsed: boolean };
-  const tearsheet = e.currentTarget as Element;
-  tearsheet
-    .querySelectorAll('[slot="header-actions"] cds-button')
-    .forEach((btn) => btn.setAttribute('size', collapsed ? 'xs' : 'sm'));
-};
-
-const toggleInfluencerPanel = () => {
-  const influencer = document.querySelector(`${prefix}-tearsheet-influencer`);
-  if (influencer) {
-    influencer.toggleAttribute('influencer-panel-open');
-  }
-};
-const toggleSummaryPanel = () => {
-  const summaryContent = document.querySelector(
-    `${prefix}-tearsheet-summary-content`
-  );
-  if (summaryContent) {
-    summaryContent.toggleAttribute('summary-panel-open');
-  }
-};
-const description = html` <c4p-truncated-text
-  slot="description"
-  lines="2"
-  type="expand"
-  expand-label="Read more"
-  collapse-label="Read less"
-  id="header-description__truncatedText"
-  value="Buttons are used to initialize an action, either in the background or foreground of an experience. There are several kinds of buttons. Primary buttons should be used for the principle call to action on the page. Secondary buttons should be used for secondary actions on each page. Danger buttons should be used for a negative action (such as Delete) on the page"
-></c4p-truncated-text>`;
-
-const progressIndicator = (vertical: boolean) =>
-  html`<cds-progress-indicator ?vertical=${vertical} current-index="1">
-    <cds-progress-step
-      description="Step 1: Getting started with Carbon Design System"
-      label="First step"
-      complete
-    ></cds-progress-step>
-    <cds-progress-step
-      description="Step 2: Getting started with Carbon Design System"
-      label="Second step with tooltip"
-      current
-    ></cds-progress-step>
-    <cds-progress-step
-      description="Step 3: Getting started with Carbon Design System"
-      label="Third step with tooltip"
-    ></cds-progress-step>
-    <cds-progress-step
-      description="Step 4: Getting started with Carbon Design System"
-      label="Fourth step"
-      secondary-label="Example invalid step"
-      invalid
-    ></cds-progress-step>
-    <cds-progress-step
-      disabled
-      description="Step 5: Getting started with Carbon Design System"
-      label="Fifth step"
-    ></cds-progress-step>
-  </cds-progress-indicator>`;
-
-const tabs = html` <cds-tabs value="tab-1">
-  <cds-tab id="tab-1" target="tab-panel-1" value="tab-1">Tab 1</cds-tab>
-  <cds-tab id="tab-2" target="tab-panel-2" value="tab-2">Tab 2</cds-tab>
-  <cds-tab id="tab-3" target="tab-panel-3" value="tab-3">Tab 3</cds-tab>
-  <cds-tab id="tab-4" target="tab-panel-4" value="tab-4">Tab 4</cds-tab>
-  <cds-tab id="tab-5" target="tab-panel-5" value="tab-5">Tab 5</cds-tab>
-  <cds-tab id="tab-6" target="tab-panel-6" value="tab-6">Tab 6</cds-tab>
-  <cds-tab id="tab-7" target="tab-panel-7" value="tab-7">Tab 7</cds-tab>
-</cds-tabs>`;
-
-const tabPanel = html` <div class="tabs-demo">
-  <div id="tab-panel-1" role="tabpanel" aria-labelledby="tab-1" hidden>
-    Tab Panel 1
-  </div>
-  <div id="tab-panel-2" role="tabpanel" aria-labelledby="tab-2" hidden>
-    Tab Panel 2
-  </div>
-  <div id="tab-panel-3" role="tabpanel" aria-labelledby="tab-3" hidden>
-    Tab Panel 3
-  </div>
-  <div id="tab-panel-4" role="tabpanel" aria-labelledby="tab-4" hidden>
-    Tab Panel 4
-  </div>
-  <div id="tab-panel-5" role="tabpanel" aria-labelledby="tab-5" hidden>
-    Tab Panel 5
-  </div>
-  <div id="tab-panel-6" role="tabpanel" aria-labelledby="tab-6" hidden>
-    Tab Panel 6
-  </div>
-  <div id="tab-panel-7" role="tabpanel" aria-labelledby="tab-7" hidden>
-    Tab Panel 7
-  </div>
-</div>`;
-
-const summaryContent = html`  <div class="rightDetailsBody">
- <Heading class="summaryPanelHeading">Summary Details</Heading>
-                <div>
-                  <label>item 1</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 2</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 3</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 4</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 5</label>
-                  <p>item description</p>
-                </div>
-              </div><div>
-                  <label>item 1</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 2</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 3</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 4</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 5</label>
-                  <p>item description</p>
-                </div>
-              </div><div>
-                  <label>item 1</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 2</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 3</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 4</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 5</label>
-                  <p>item description</p>
-                </div>
-              </div><div>
-                  <label>item 1</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 2</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 3</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 4</label>
-                  <p>item description</p>
-                </div>
-                <div>
-                  <label>item 5</label>
-                  <p>item description</p>
-                </div>
-              </div>`;
-
-const dummyContent = html` <section class="main-content">
-  <h4>Main content heading</h4>
-
-  <div class="${storyPrefix}text-inputs">
-    <cds-text-input
-      label="Enter an important value here"
-      id="input1"
-    ></cds-text-input>
-    <cds-text-input
-      label="Here is an entry field:"
-      id="tss-ft2"
-    ></cds-text-input>
-  </div>
-
-  <div class="${storyPrefix}text-inputs">
-    <cds-text-input
-      label="Enter an important value here"
-      id="tss-ft1"
-    ></cds-text-input>
-    <cds-text-input
-      label="Here is an entry field:"
-      id="tss-ft2-2"
-    ></cds-text-input>
-  </div>
-
-  <div class="${storyPrefix}text-inputs">
-    <cds-text-input
-      label="Enter an important value here"
-      id="tss-ft3"
-    ></cds-text-input>
-    <cds-text-input
-      label="Here is an entry field:"
-      id="tss-ft4"
-    ></cds-text-input>
-  </div>
-
-  <div class="${storyPrefix}textarea-container">
-    <cds-textarea label="Notes" value="This is a text area"></cds-textarea>
-    <cds-textarea label="Notes" value="This is a text area"></cds-textarea>
-    <cds-textarea label="Notes" value="This is a text area"></cds-textarea>
-    <cds-textarea label="Notes" value="This is a text area"></cds-textarea>
-  </div>
-</section>`;
+import {
+  storyPrefix,
+  getButtonSize,
+  getDecorator,
+  toggleButton,
+  handleCollapseChange,
+  toggleInfluencerPanel,
+  toggleSummaryPanel,
+  description,
+  progressIndicator,
+  tabs,
+  tabPanel,
+  summaryContent,
+  dummyContent,
+  Bee,
+  Add16,
+  RightPanelClose32,
+  iconLoader,
+} from './tearsheet-helpers';
 
 export const Default = {
   args: {
